Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Thanks for the advice, Alan. This is taking me a bit longer because I was stuck on named types not working properly. I think there is a bug in go/types now; will get to that first.

Well, I don't know if there is a bug in go/types or not, but what happens is:

type A foo.B

A.Underlying() == foo.B.Underlying()

I've read the go/types source, and for a long time it has "squashed" underlying named types.

With that rule, it seems to me like it's impossible for vet to know where the non-named underlying type of a type was declared. If we get to a types.Struct, it's too late - it has no object nor position. And if we get a types.Named, there will only be one - the first one encountered.

What should we do here? Am I using go/types wrong?

Patch-set: 2
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>