go/types: prevent errors involving invalid types
This could happen under certain circumstances, like:

	var x *missingType
	var _ = x.method

Here, x's type is *Pointer, with an invalid type as its element type.
Many pieces of go/types check if a type is valid, for example to decide
whether or not to report "has no field or method" errors.

The result is a confusing error:

	x (variable of type *invalid type) has no field or method ..

To fix this, add an isInvalid method to the operand type. On top of
checking if the operand has the invalid mode, it checks if any part of
its type is invalid. For now, just use it in the selector func, where
the issue was found.

Fixes #24182.

Change-Id: Idc447948781393d47da29db9178cb6abd5000fe8