go/types: prevent Type.Elem() == Typ[Invalid]
This could happen under certain circumstances, like:

	var x *missingType
	var _ = x.method

Here, x's type is *Pointer, with an invalid type as its element type.
Many pieces of go/types check if a type is valid, for example to decide
whether or not to report "has no field or method" errors.

The result is a confusing error:

	x (variable of type *invalid type) has no field or method ..

To fix this, prevent such broken types from being returned by typExpr.
Don't do this for named types with invalid underlying types, as those do
serve a function - they mean that the type or expression has an issue,
but it has not been reported yet.

Fixes #24182.

Change-Id: Idc447948781393d47da29db9178cb6abd5000fe8