cmd/compile,test: combine byte loads and stores on ppc64le
CL 74410 added rules to combine consecutive byte loads and
stores when the byte order was little endian for ppc64le. This
is the corresponding change for bytes that are in big endian order.
These rules are all intended for a little endian target arch.

This adds new testcases in test/codegen/memcombine.go

Also fixes a regression in the little endian 16 bit stores.

Fixes #22496
Updates #24242

Change-Id: Idb3bd1e6e7b12d86cd828fb29cb095848a3e485a