Update patch set 1
Patch Set 1:

The implementation looks pretty reasonable, but this is a very large increase to the API surface of the runtime, and encodes decisions like how many GCs to report in that API. AFAIK the runtime package is covered under the Go 1.0 compatibility promise, which means that those decisions could never be changed. I think we need to do better, but I'll leave it up to Rick and maybe Austin to decide.

Patch-set: 1
CC: Gerrit User 5167 <5167@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 17090 <17090@62eb7196-b449-3ce5-99f1-c037f21e1705>