Update patch set 3
Patch Set 3:

> Does this need to go through the proposal process for the new API?
> 
> The new API should probably be highlighted more in the commit message, though.
> 
> I'd make the subject say which symbol is being added.

I haven't looked through the CL yet.

If this follows the API that Russ outlined at #9661 then I don't think this needs to go through the proposal process.  It's just a small extension of the existing RawConn support.

I agree that the new API should be in the commit message.

Patch-set: 3