Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Unless I misunderstand this badly, this seems very likely to break existing code.

Ian, if this solution in line with the solution outlined in https://github.com/golang/go/issues/21092 can you comment there with your concerns?

Otherwise, what is it about this change that is different from that issue’s proposed fix and how would it break existing code?

Patch-set: 2