Update patch set 2
Patch Set 2:

Hello Andrew and Ian,

First, thanks for chiming in, I sincerely appreciate it. This is my first (potential) contribution to Golang, so thanks for working with me on this.

I should give some more context to this PR, so please bear with me as it is sort of long.

First, Ian you are correct, this PR does break existing behavior. In fact, if you pull down and apply this code, you will see a few broken tests! The tests I believe, are merely documenting existing behavior, not desired behavior, which is why I left them untouched. I will most certainly write/edit more as we continue.

The key issue that this PR attempts to solve is the desired behavior for handling pre-existing data of a struct when written to by json.Unmarshal.  What is that desired behavior? That is the real debate. Currently, data from a 'reference' type (slice or map in this case) has the potential to bleed in between successive unmarshallings if the same target is used (i.e., unmarshalling a slice of DB query results in a loop, reusing the same target).

My colleagues and I believe that the existing struct should be zero'ed out prior to deserialization. From our prior experience with a multitude of different languages, we thought this was a reasonable assumption. That assumption was incorrect however in the case of Golang's encoding/json implementation, and was the root of a bug in our code.

From our perspective, if we unmarshal JSON (either from a NoSQL DB query result, a settings file, etc), we expect the target struct to represent that JSON object exactly,regardless of wether the target variable was initialized or not.

We are not the only programmers in the world, our perspectives are only one of a near infinite amount, so we wanted to open this PR, and a discussion, on what should be happening (hopefully with a contribution to resolve it as well!)

Patch-set: 2
CC: Gerrit User 26807 <26807@62eb7196-b449-3ce5-99f1-c037f21e1705>