Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I haven't tested it, but from reading it, this CL will break the program that Russ links to in https://github.com/golang/go/issues/21092#issuecomment-346231694 .

Ian,

Yes, it does break that example. As stated above, it just seems like an odd user experience to unmarshal JSON and have the resulting struct not match that JSON exactly.

If there is some use case where a blend of data is required, it should be done outside of the Unmarshal function for both readability and maintainability.

Patch-set: 2