Update patch set 1
Patch Set 1:

We use `nil` all over the place in that directory.  It's defined in libcgo.h.  What if we edit the files to move the #include of the CoreFoundation files before the #include of libcgo.h?  We might also need #undef nil in libcgo.h.

Or we could change the entire directory to use NULL instead of nil.  I'm OK with that too.  But I don't want to just change a few uses of nil.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>