Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (2 comments)
> 
> Looks like the PPC64.rules file generates a much larger rewrite file than AMD64 or ARM64 even though our rules file is smaller. I'll have to investigate why and see if there's anything I can do to improve it.

Not sure if you saw my post in golang-dev. Our rewritePPC64.go file is excessively large because rulegen generates all combinations of the operand orderings for a rule when the operation is commutative, so for the 8 byte rules using OR, a huge number of match cases are generated. Apparently for ARM64, many of the rules used for this case are not commutative so they don't generate so much code. In AMD64, they hit this problem as well but changed their rules to be smaller by reusing previous rules. I think they may be considering a way to resolve this. I will spend some time trying to see if I could reuse rules to reduce the size.

Patch-set: 6
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>