Update patch set 3
Patch Set 3:

> With a very quick look, is it due to that FP != callerSP on LR machines? (there is fixed frame space for the LR slot.)

That seems possible.

It looks like in runtime.argp, we just take the address of the first parameter, which matches what the compiler used to do. I thought this was equivalent to getcallersp, but it seems I thought wrong.

It seems like we probably have two options:

1) Change argp to use "return getcallersp(unsafe.Pointer(&x))"

2) Revert the callerSP change in the compiler.

I'd slightly favor the former, just because there's more established convention around using callerSP. But if there are any technical reasons to favor the latter, I'm fine with that too.

Patch-set: 3