Update patch set 1
Patch Set 1: Code-Review-1

Hrm, this works for racewalk because we call simply:

    racefuncenter(getcallerpc())

so literally all we need is

    CALL getcallerpc
    CALL racefuncenter

But ssa.go is too late for doing scheduling of assigning to parameter slots on the stack. For example, f(42, getcallerpc()) gets compiled to something like:

    MOV $42, 0(SP)
    CALL getcallerpc
    MOV 0(SP), R1
    MOV R1, 8(SP)
    CALL f

So this happens to work today, but would break if we try to use it for anything else. :/

Patch-set: 1
Reviewer: Gerrit User 5440 <5440@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1