cmd/compile: enable -dwarflocationlists by default
Fixes a bug and a follow-on bug, and enables dwarf location
lists.  This should also substantially improve the debugging
experience in the location lists enabled case by providing
debugging information for more variables.

Dwarf locationlists are (only) emitted for values that
appear in SSA.  "Values" that are not eligible for SSA
transformation, either because their addresses are taken or
because they are too large for the decomposition heuristic,
did not appear at all or we treated as "unavailable".
Including those values also meant that in some cases
NAME_DELETED_AUTO is required.

To fix, gc/ssa.go records non-temp autos that are ineligible
for ssa.  The old debuginfo function was replaced with a
closure including that list, passed to *Progs.Flush, then
obj.FlushPlist, then to to obj.populateDWARF. The debugInfo
function stored in Ctxt is not used because populateDWARF
may be called from worker threads operating in parallel.

In the modified debugInfo, non-SSA autos are processed in
the same way that they are when location lists are not
present, except that NAME_DELETED_AUTO is also emitted
because of the possibility of their removal by dead code
elimination.

(Finally) the default value for the -dwarflocationlists
flag is set to true, and ssa/debug_test.go has the flag
removed since it is now on by default.

Change-Id: I1b420bb2b07033589d3e90c5a20fe069ca12d649