Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> (2 comments)
> 
> Thanks for finding this, and sorry that the code is apparently tricky to understand :(

We suspected this might be the problem, but wanted you to have a look.
Thanks.  Will be interesting to see if this undoes the gdb-testing failure in the turn-on CL.
(I think that's disabled on OSX now because of default gdb unavailability.)

Patch-set: 2
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>