sync: use a single global pool instead of per-P shared ones
Rework how sync.Pool is structured internally: instead of per-P private and shared
pools, have a per-P private pool and a global shared pool.

The per-P private pool is now composed of one fixed and one additional shard (each
shard contains 32 objects). When the additional shard is full, it gets moved to
the global pool. When the additional shard is empty, a full shard is moved from
the global pool to the private pool. Moving shards is extremely cheap because it
is effectively a single pointer-sized write.

To control access to the shared pool a non-blocking mutex is used (trylock). Because
the private pools are much bigger than previously, accesses to the global pools are
less frequent (at least in non-pathological cases, but even in pathological cases
like the one in the PoolUnderflowUnbalanced benchmark this approach is faster).

The non-blocking mutex approach is faster but can lead to increased memory usage
since contention on the global pools can cause some Get() calls to return nil even
if objects exist in the global pools. Similarly, it can cause some Put() calls to
fail to add the object to the pool.

Another source of increased memory consumption is the increase in size of the per-P
private pool, that goes from 1 to 32 elements. The shardSize (32) is somewhat
arbitrary and can be further tweaked to trade off memory usage for CPU usage.

To minimize shard allocations a second global pool (globalFree) is used as a
freelist for empty shards. Both global pools are implemented as a simple linked
list that is used as a stack.

On my MacBook Pro this CL has the following effects:

name                       old time/op  new time/op  delta
Pool-4                     14.6ns ± 4%   9.0ns ± 2%  -38.06%  (p=0.000 n=18+19)
PoolOverflow-4             1.87µs ± 2%  1.06µs ± 1%  -43.56%  (p=0.000 n=18+20)
PoolUnderflowUnbalanced-4   151ns ± 6%    39ns ± 1%  -73.99%  (p=0.000 n=20+20)
PoolOverflowUnbalanced-4   64.9ns ± 8%  17.5ns ±10%  -73.12%  (p=0.000 n=20+19)

Benchmark results on machines with higher core counts would be very welcome.

This CL supersedes CL 95579 and is a (likely better) alternative to CL 49110.

Potentially related to issue #22950.

Change-Id: Ic16e2954a4302385fbc6f165d468600adc608069