Update patch set 1
Patch Set 1:

(3 comments)

> Patch Set 1:
> 
> It's ok to merge BIC into AND. But BIC is already used by existing code as shown below:
> ./runtime/asm_arm64.s:  BIC     $0xf, R4, R6;                   \
> ./internal/bytealg/equal_arm64.s:       BIC     $0x3f, R1, R3
> ./internal/bytealg/equal_arm64.s:       BIC     $0xf, R1, R3
> ./internal/bytealg/count_arm64.s:       BIC     $0x1f, R0, R3
> ./internal/bytealg/count_arm64.s:       BIC     $0x1f, R2, R9
> ./internal/bytealg/indexbyte_arm64.s:   BIC     $0x1f, R0, R3
> 
> Do you plan to change them?

No, I don't plan to change them. To be clear, "merge the optab entries of BIC to AND" is just a smaller refactoring of the code, not a user-facing change. The assembly language and what bytes it generates are unchanged for this part.

Patch-set: 1
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>