Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (2 comments)

Yes, that is exactly what I mean. I would use LDAR as the entry in the Optab and then add the others in the case in buildop where they call opset.

The fact that the ECIWX/ECOWX are not even in ISA 3.0 is another good reason to separate them.

Patch-set: 2
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>