Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> (1 comment)
> 
> This CL probably needs a test. Is there a corresponding issue? If so, add
> 
> Fixes #xxxx
> 
> to the commit message.

There's an issue on GitHub for this, but is that what you mean here? If not, what should I do? I could write a test that fails without this patch and succeeds with it (there's a reason I wrote this after all) but is there some instruction on how to do that?

Patch-set: 3
CC: Gerrit User 7435 <7435@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 25894 <25894@62eb7196-b449-3ce5-99f1-c037f21e1705>