cmd/compile: optimize struct partial self-assign
Instead of explicitly zeroing a subset of fields
to reset and keep all other fields it is usually
better to explicitly copy a subset of fields to
preserve and zero the rest implicitly.

Ungortunately, this second approach would result
in inefficient code being generated. Detect this
pattern and convert it to an equivalent assignment.

No significant impact on go1 benchmarks on darwin/amd64

Fixes #24386

Change-Id: I5543ab33cf0a50dc941287790beb0209efb3eacc