cmd/compile: optimize struct partial re-assign
Instead of explicitly zeroing a subset of fields
to reset and keep all other fields it is usually
better to explicitly copy a subset of fields to
preserve and zero the rest implicitly.

Unfortunately, this second approach would result
in inefficient code being generated. Detect this
pattern and convert it to an equivalent assignment.

For large structures, avoid ballooning assignments
by coalescing adjacent fields into a single call
to memclr if appropriate. Care is taken to invoke
the write barrier as needed.

To maximize coalescing opportunities and reduce
call overhead, the bulk write barrier may be called
directly instead of through memclHasHeapPointers.

No significant impact on go1 benchmarks on darwin/amd64

Fixes #24386

Change-Id: I5543ab33cf0a50dc941287790beb0209efb3eacc