Update patch set 1
Patch Set 1:

> > Note that committing this will almost certainly break #20703.
> 
> I wanted to verify this and ran the tests on an s390x gomote after reverting ec7c6c16e14a46d907c7dc95662093f0ed846143 (which is essentially what this CL is doing).
> 
> I could not reproduce #20703 — under strace, we never exceed 100 file descriptors.
> 
> iant, could you test whether you can still reproduce the issue on s390x? Perhaps something else changed.

There is nothing special about s390 here.  It just occurred on s390 because it was using a low default ulimit.  The other characteristics required are a relatively slow connection to the DNS server--the s390 was using 8.8.8.8--and having the sem channel in TestTCPSpuriousConnSetupCompletionWithCancel be larger than the number of goroutines, so that some of them stall out.  I can still reproduce the issue on my laptop if I make those changes.

That said, I do see now that the earlier change (CL 45999) is not right.  We should not be memoizing the result of a context cancellation.  And simply reverting CL 45999 isn't enough.  After all, another DNS lookup could arrive between the context cancellation and the lookupGroup.Forget, could get folded into the existing singlegroup lookup, and could see the cancellation.

Patch-set: 1
Reviewer: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>