cmd/compile: enable scopes unconditionally
This revives Alessandro Arzilli's CL to enable scope
whenever any dwarf is emitted (with optimization or not),
adds a test that detects this changes and shows that it
creates more truthful debugging output.

Also included are updates to the Delve test output (it
had fallen out of sync; creating test output for one
updates it for all) and minor naming changes in
ssa/debug_test.

Benchmarking to come.

Change-Id: Id53b8d57bd00440142ccbd39b95710e14e083fb5