Update patch set 2
Patch Set 2:

(3 comments)

Thanks for providing a asymptotically better algorithm!

As this algorithm is much more complex as SymMerge
(a factor of 4 in line count) it should be demonstratible
faster for actual input sizes. As explained in a reply
to a comment I would propose to add realistic and large
benchmarks first.

Patch-set: 2