regexp: use sync.Pool to cache regexp.machine objects
Performance optimization for the internals of the Regexp type. This adds
no features and has no user-visible impact beyond performance.

The before v. after benchmark results show a lot of noise from run to
run, and since none of the benchmarks are concurrent, I don't expect any
difference. The impact to my load test of the Prometheus code base shows
that the lock contention in Regexp is gone, and the server is now able
to scale linearly 2.5x times more than before (and I only stopped there
because I ran out of CPU in my test machine).

Fixes #24411

Change-Id: Ib33abff2802f27599f5d09084775e95b54e3e1d7