Update patch set 1
Patch Set 1:

I don't know how to maintain packages in the vendor directory. There is a vendor.json but from the history, I found it's not consistently updated and became out-of-sync. I created this change with manual copy (cp -r) and manual edit. Let me know if that's not the right process.

Would be nice if these are migrated to the new vgo-way.

Patch-set: 1
CC: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>
Reviewer: Gerrit User 5190 <5190@62eb7196-b449-3ce5-99f1-c037f21e1705>