Update patch set 2
Patch Set 2:

> I seem to recall this same change being rejected in the past. Somebody might want to search to remember what the concern was.

Yes.  A similar change was made in https://golang.org/cl/44150043, by you.  It was rolled back in  https://golang.org/cl/48190043.  Dmitry said: "The remaining concern is: Are there cases where a program creates lots of local short-lived regexps (one-shot), and it's not possible to replace them with global ones? For this case we've introduced a global contended mutex."  Based on that comment, the CL was rolled back until there was a clear performance win.

Patch-set: 2