cmd/internal/obj/arm64: fix the bug of incorrect handling negative offset of LDP
The current assembler will report error when the LDP negative offset
is in the range of [-256, 0) and is not the multiples of 8.

The fix introduces C_NSAUTO_8 and makes C_NPAUTO includes it instead of
C_NSAUTO, so that assembler will treat the negative offset that is greater
than -256 and is the multiples of 8 as C_NSAUTO_8, and let it enter the
C_NPAUTO entry. The fix also lets the offset mentioned above enter the
C_LAUTO entry.

Add the test cases.

Fixed #24471

Change-Id: I42f34e3b8a9fc52c9e8b41504294271aafade639