cmd/vet: only zero/non-zero exit status matters
"go vet" exits with 2 on type-checking errors (because the compiler
itself does so), but with 1 on vet errors.

But we're really not trying to convey any nuance in vet's exit status.
It's really just one bit: success or failure. Update the documentation
so it doesn't claim anything subtler than this.

Fixes #24477.

Change-Id: I52333f072f4ca6fbc6e5e4483c37cb5aa941aa60