Update patch set 3
Patch Set 3:

> I was not aware there was a concern about adding new assembly. This one was easy to do and it had assembly for other platforms so I assumed someone thought it was important.

Maybe we should just delete all the assembly for rc4... I did a quick smoke test on my laptop (amd64) and the pure Go version at tip seems to be, if anything, a little bit faster that the assembly. Given that rc4 is cryptographically broken I don't see any security reasons to keep the assembly. I thought about adding s390x assembly for this at one stage but it didn't seem worthwhile.

Lynn: would it be possible to teach the compiler some of the tricks you are using in the assembly for this to get the performance closer? The assembly here seems like it might make a good 'goal' for new compiler optimizations.

Patch-set: 3
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>