Update patch set 3
Patch Set 3:

In general we prefer using Go to write Go instead of C and instead of assembly. The less assembly we have, the better. We view it as a necessary evil in some cases where performance really matters. And yes, we should make the compiler do the tricks when possible. Adding more assembly increases the (perceived?) cost of adding a new GOARCH for other people, who feel compelled to match assembly parity (as you've done with this CL).

I'd argue that md5 is still important. It's increasingly used as a low-security checksum (e.g. the AWS S3 API, which others have copied). 

But rc4 seems esoteric.

Patch-set: 3