Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> In general we prefer using Go to write Go instead of C and instead of assembly. The less assembly we have, the better. We view it as a necessary evil in some cases where performance really matters. And yes, we should make the compiler do the tricks when possible. Adding more assembly increases the (perceived?) cost of adding a new GOARCH for other people, who feel compelled to match assembly parity (as you've done with this CL).
> 
> I'd argue that md5 is still important. It's increasingly used as a low-security checksum (e.g. the AWS S3 API, which others have copied). 
> 
> But rc4 seems esoteric.

I'm OK with withdrawing it, but it does indicate that there is a lot of room for improvement in Go's code generation for ppc64x. And I don't think it is a matter of adding a few simple tricks, it would involve improved loop optimizations, register allocator, and better scheduling for ppc64x.

Patch-set: 3
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>