Update patch set 8
Patch Set 8:

Thanks for putting a lot of effort in implementing an algorithm
with better asymptotical complexity. I really would like to see
this being merged as the current Stable is suboptimal for large
inputs.

Unfortunately the asymptotic complexity of an algorithm is pretty
uninteresting if the constants hidden in the big-O makes it
slower than a asymptotically worse algorithm. There's a practical
reason why Knuth gives these factors (for a certain machine).
Any new algorithm for Stable should be faster, at least for some
inputs sizes or input types.
I experimented with your CL and could not reproduce a speedup
for input sizes from 1e1 to 1e6 (some benchmarks where slower
than the original SymMerge based Stable) just like your own
benchmarks show.

The existing Stable uses SymMerge making O(n*log(n)*log(n))
calls to data.Swap while the new Stable uses
StableOptimalBlockMerge (SOBMerge) making only  O(n*log(n))
calls to data.Swap. SymMerge is very simple (measured in LoC)
while SOBMerge is pretty complex (again in LoC).
From these facts my gut feeling would be that SymMerge will
beat SOBMerge below a certain input size n < threshold, just
like InsertionSort beats QuickSort, Stable, etc, for small n.
So I would expect a runtime-optimal algorithm for stable sorting
to use three different strategies:
   InsertionSort for  n < threshold_1
   SymMerge    for threshold_1 <= n < threshold_2
   SOBMerge    for threshold 2 <= n
Especially as the difference between SymMerge and SOBMerge
is "only" log(n).

One fundamental question is: Is SOBMerge the right algorithm
to use in stable sorting _at_ _all_?
SOBMerge is interesting and novel, but its focus is on merging
blocks of _different_ size (which is an interesting question by itself).
But the sizes of the blocks to merge in our case are
under _our_ _own_ _control_.
We do not need an algorithm which is optimal for a given k:
We can _choose_ k so that our algorithm is optimal.
That's why I am sceptical that SOBMerge is the best algo for
our problem here. If we are willing to introduce a movement
imitation buffer then maybe a different algorithm might be
better (in the sense of faster for real inputs).

Nevertheless it seems faster for very large input size making
it at least a candidate for merging large blocks.
Please pay attention to how SOBMerge behaves on pathologic
inputs: Already sorted, sorted in reverse, all constant input,
almost constant input, just a few values. The stuff that 
testBentleyMcIlroy tests and counts. While playing around
with your code I got the impression, that SOBMerge is not
too well suited to such pathologic input.
(I think Kim & Kutzner tested SOBMerge on random input
data; unfortunately the are not explicit in their paper. So take
their claim that SOBMerge is "the fastest" with a grain of salt.)
Please include the new, larger Benchmarks in the patch, so
that we know on what type of input SOBMerge is faster.

Patch-set: 8