cmd/compile: detect simple inductive facts in prove
Currently, we compile range loops into for loops with the obvious
initialization and update of the index variable. In this form, the
prove pass can see that the body is dominated by an i < len condition,
and the loopbce pass can further detect that i is an induction
variable and that 0 <= i < len.

GOEXPERIMENT=preemptibleloops compiles range loops to OFORUNTIL and
we're preparing to unconditionally switch to a variation of this for
 #24543. OFORUNTIL moves the increment and condition *after* the body,
which makes the bounds on the index variable much less obvious. With
OFORUNTIL, proving anything about the index variable requires
understanding the phi that joins the index values at the top of the
loop body block.

This interferes with both prove's ability to see that i < len (this is
true on both paths that enter the body, but from two different
conditional checks) and with loopbce's ability to detect the induction
pattern.

Fix this by teaching prove to detect that the index in the pattern
constructed by OFORUNTIL is an induction variable and add both bounds
to the facts table. This is generally nicer than trying to fix loopbce
since prove can use these facts for much more than eliminating slice
indexes. This isn't powerful enough to replace loopbce, which still
detects many, many other induction variables, though it might not be a
bad start to replacing loopbce.

Without any GOEXPERIMENT, this fires 8 times for std: 7 of these were
previously detected by loopbce and are now handled by prove, plus 1
extra that was never handled by prove or loopbce. With
GOEXPERIMENT=preemptibleloops, we previously failed to prove 30 bounds
checks we were able to prove before; with this change we fail to prove
just 1 and eliminate several bounds checks we couldn't before.

Change-Id: Ic025d669f81b53426309da5a6e8010e5ccaf4f49